#133: Add DbConfig helper
